refactor(core): split mark style from within-mark paint
mark-paint.ts mixed two edit axes: renderer-neutral shape/dash/stroke
resolution and #591 gradient/glow paint. Tests already mirrored the
split; extract mark-style.ts and keep paint-only symbols in mark-paint.

Public @ggsvelte/core export names are unchanged.
